ar0521-dts: Disable media-control
media-control was enabled to make driver compatible with
libcamera, but we are seeing some issue with
gstreamer v4l2src when media-control is enabled.

Since, libcamera is not in use right now, disable media control
untill gstreamer v4l2src is compatible with media-control or when
there is plan to move on libcamera or gstreamer with libcamsrc.

Signed-off-by: Tejas Patel <tejaspatel.12@outlook.com>